+++ src/components/HeaderBar/HeaderBurgerMenu/HeaderMenuItem/HeaderMenuItem.mdx
@@ -28,7 +28,7 @@ Ce composant doit être utilisé dans un composant `HeaderMenuSection` afin de r
 `}
 />
 
-Pour plus de détails sur le contexte d'utilisation de ce composant, veuillez consulter le composant `HeaderComplexMenu`.
+Pour plus de détails sur le contexte d'utilisation de ce composant, veuillez consulter le composant `HeaderBurgerMenu`.
 
 ## Usage
