+++ src/components/HeaderBar/HeaderBurgerMenu/HeaderSubMenu/HeaderSubMenu.mdx
@@ -5,7 +5,7 @@ import * as HeaderSubMenuStories from "./HeaderSubMenu.stories";
 
 # HeaderSubMenu
 
-Le composant `HeaderSubMenu` est un sous-composant du composant `HeaderComplexMenu` qui permet de créer un sous-menu dans le menu principal.
+Le composant `HeaderSubMenu` est un sous-composant du composant `HeaderBurgerMenu` qui permet de créer un sous-menu dans le menu principal.
 En mode mobile, les sous-menus s'affichent par dessus du menu parent, en mode desktop, les sous-menus s'affichent à droite du parent.
 Les sous-menus peuvent être imbriqués pour créer des sous-sous-menus.
