+++ src/components/HeaderNavbar/HeaderNavbar.mdx
@@ -7,7 +7,7 @@ import * as HeaderNavbarStories from "./HeaderNavbar.stories";
 
 Le composant `HeaderNavbar` est un composant de navigation qui permet de naviguer entre les différentes pages de l'application.
 
-Il est basé sur les composants `HeaderBar` et `HeaderComplexMenu` du design system.
+Il est basé sur les composants `HeaderBar` et `HeaderBurgerMenu` du design system.
 
 Ce composant propose un navigation à l'aide d'un menu horizontal en mode desktop ce qui peut être util pour les produits ayant peut de page a destination des agents.
 Il est conseillé de ne pas utiliser le menu horizontal pour afficher plus de 6 liens. Vous pouvez configurer le nombre de lien affiché dans le menu horizontal en utilisant la prop `maxHorizontalMenuItems`. Au dela de ce chiffre, le menu passera en mode burger. Par defaut, le menu horizontal affiche au maximum six items.
@@ -89,10 +89,10 @@ Le slot `navigation-bar-content` quand a lui, remplace les items du menu tel que
 
 #### Les slots concernant le menu burger
 
-Le menu burger est généré automatiquement à partir des items passés en props, il utilise en interne le composant `HeaderComplexMenu`.
+Le menu burger est généré automatiquement à partir des items passés en props, il utilise en interne le composant `HeaderBurgerMenu`.
 
 
 les slots `burger-menu-prepend` et `burger-menu-append` permettent d'ajouter du contenu en haut et en bas du menu burger.
 
 Le slot `burger-menu-content` permet de remplacer le contenu du menu burger. Il n'écrase pas les slots `burger-menu-prepend` et `burger-menu-append` et permet d'utiliser les composants `HeaderMenuItem` et `HeaderMenuGroup` pour construire le menu.
-Pour plus de d'information sur la manière de construire un menu avec le composant `HeaderComplexMenu`, vous pouvez consulter la documentation de ce composant.
+Pour plus de d'information sur la manière de construire un menu avec le composant `HeaderBurgerMenu`, vous pouvez consulter la documentation de ce composant.
