+++ src/components/index.ts
@@ -7,10 +7,10 @@ export { default as DownloadBtn } from './DownloadBtn/DownloadBtn.vue'
 export { default as FooterBar } from './FooterBar/FooterBar.vue'
 export { default as FranceConnectBtn } from './FranceConnectBtn/FranceConnectBtn.vue'
 export { default as HeaderBar } from './HeaderBar/HeaderBar.vue'
-export { default as HeaderComplexMenu } from './HeaderBar/HeaderComplexMenu/HeaderComplexMenu.vue'
-export { default as HeaderMenuItem } from './HeaderBar/HeaderComplexMenu/HeaderMenuItem/HeaderMenuItem.vue'
-export { default as HeaderMenuSection } from './HeaderBar/HeaderComplexMenu/HeaderMenuSection/HeaderMenuSection.vue'
-export { default as HeaderSubMenu } from './HeaderBar/HeaderComplexMenu/HeaderSubMenu/HeaderSubMenu.vue'
+export { default as HeaderBurgerMenu } from './HeaderBar/HeaderBurgerMenu/HeaderBurgerMenu.vue'
+export { default as HeaderMenuItem } from './HeaderBar/HeaderBurgerMenu/HeaderMenuItem/HeaderMenuItem.vue'
+export { default as HeaderMenuSection } from './HeaderBar/HeaderBurgerMenu/HeaderMenuSection/HeaderMenuSection.vue'
+export { default as HeaderSubMenu } from './HeaderBar/HeaderBurgerMenu/HeaderSubMenu/HeaderSubMenu.vue'
 export { default as HeaderLogo } from './HeaderBar/HeaderLogo/HeaderLogo.vue'
 export { default as HeaderMenuBtn } from './HeaderBar/HeaderMenuBtn/HeaderMenuBtn.vue'
 export { default as HeaderNavbar } from './HeaderNavbar/HeaderNavbar.vue'
